+++ src/components/Air/Gauge.js
