+++ src/components/City/SensorsQualityStats/QualityStats.js
