+++ src/pages/air/index.js
@@ -7,10 +7,10 @@ import HowSensorsWork from './HowSensorsWork';
 import City from './City';
 import JoinNetwork from './JoinNetwork';
 import Navbar from '../../components/Header/Navbar';
-import AirHeader from '../../components/AirComponents/AirHeader';
+import AirHeader from '../../components/Air/AirHeader';
 import Footer from '../../components/Footer';
-import IndoorOutdoor from '../../components/AirComponents/IndoorOutdoor';
-import Issues from '../../components/AirComponents/Issues';
+import IndoorOutdoor from '../../components/Air/IndoorOutdoor';
+import Issues from '../../components/Air/Issues';
 import Stories from '../../components/About/Stories';
 import Support from '../../components/Support';
 import HealthClimate from './HealthClimate';
